Casts for disabling compiler warnings included.
git-svn-id: file:///Users/min/Documents/svnrepros/svn_mindia/trunk@69 5f0a215f-9dc6-43d0-aa08-c0b9dc37ed8b

+++ src/misctools.cpp
@@ -248,7 +248,7 @@ bool IsImageFileDrag( const QMimeSource * pEvent )
 	if( QUriDrag::decode( pEvent, aStrList ) )
 	{
 		// all droped file-names have to be images with known format !!!
-		for( int i=0; i<aStrList.count(); i++ )
+		for( int i=0; i<(int)aStrList.count(); i++ )
 		{
 			const QString sTest = aStrList.at( i );
 			const char * s = (const char *)sTest;
@@ -293,7 +293,7 @@ bool IsSoundFileDrag( const QMimeSource * pEvent )
 	if( QUriDrag::decode( pEvent, aStrList ) )
 	{
 		// all droped file-names have to be images with known format !!!
-		for( int i=0; i<aStrList.count(); i++ )
+		for( int i=0; i<(int)aStrList.count(); i++ )
 		{
 			const QString sTest = aStrList.at( i );
 			const char * s = (const char *)sTest;
